refactor(web): refactor cost unit settings
