+++ apps/web/src/modules/settings/components/org-settings/index.ts
@@ -1,1 +1,1 @@
-export { OrgSettingsCostUnits } from "./org-settings-cost-units";
+export { OrgSettingsCostUnits } from "./cost-units";
