+++ apps/web/src/modules/settings/components/org-settings/update-cost-unit.tsx
@@ -1,14 +1,28 @@
 "use client";
 
-import { Dialog as DialogPrimitive } from "@base-ui/react";
+import {
+	AlertDialog as AlertDialogPrimitive,
+	Dialog as DialogPrimitive,
+} from "@base-ui/react";
 import { useForm } from "@tanstack/react-form";
 import { useSuspenseQueries } from "@tanstack/react-query";
 import type { CostUnitGroup } from "@zemio/db";
-import { InfoIcon } from "lucide-react";
+import { InfoIcon, TrashIcon } from "lucide-react";
 import React from "react";
 import { toast } from "sonner";
 import type z from "zod";
 import { AsyncBoundary } from "@/components/async-boundary";
+import {
+	AlertDialog,
+	AlertDialogAction,
+	AlertDialogCancel,
+	AlertDialogContent,
+	AlertDialogDescription,
+	AlertDialogFooter,
+	AlertDialogHeader,
+	AlertDialogTitle,
+	AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 import {
 	Field,
@@ -18,6 +32,7 @@ import {
 	FieldLabel,
 } from "@/components/ui/field";
 import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
 import {
 	NativeSelect,
 	NativeSelectOptGroup,
@@ -32,125 +47,159 @@ import {
 	SheetHeader,
 	SheetTitle,
 } from "@/components/ui/sheet";
-import { Skeleton } from "@/components/ui/skeleton";
 import {
 	Tooltip,
 	TooltipContent,
 	TooltipTrigger,
 } from "@/components/ui/tooltip";
-import type { WithHandle } from "@/lib/types";
-import { createCostUnitSchema } from "@/lib/validators";
+import { NO_COST_UNIT_GROUP } from "@/lib/consts";
+import { updateCostUnitSchema } from "@/lib/validators";
 import { api } from "@/trpc/react";
-import { ExamplesInput } from "./org-settings-cost-units";
-import {
-	type CreateCostUnitGroupHandle,
-	CreateCostUnitGroupSheet,
-} from "./org-settings-create-cost-unit-group";
+import { ExamplesInput } from "./examples-input";
+import { SheetFormError, SheetFormSkeleton } from "./sheet-form-state";
 
-type CreateCostUnitFormValues = z.infer<typeof createCostUnitSchema>;
-type CreateCostUnitHandle = ReturnType<typeof DialogPrimitive.createHandle>;
+type UpdateCostUnitPayload = { id: string } | undefined;
+type UpdateCostUnitHandle = ReturnType<
+	typeof DialogPrimitive.createHandle<UpdateCostUnitPayload>
+>;
+type UpdateCostUnitFormValues = z.infer<typeof updateCostUnitSchema>;
 
-const FORM_ID = "create-cost-unit";
-const NO_COST_UNIT_GROUP = "NO_GROUP" as const;
+const FORM_ID = "update-cost-unit";
+const COST_UNIT_FORM_FIELD_COUNT = 4;
 
-function createCreateCostUnitHandle(): CreateCostUnitHandle {
-	return DialogPrimitive.createHandle();
+function createCostUnitUpdateHandle() {
+	return DialogPrimitive.createHandle<UpdateCostUnitPayload>();
 }
 
-function CreateCostUnitSheetTrigger({
+function UpdateCostUnitSheet({
 	handle,
 	...props
-}: React.ComponentProps<typeof Button> & WithHandle) {
+}: Omit<React.ComponentProps<typeof Sheet>, "handle"> & {
+	handle: UpdateCostUnitHandle;
+}) {
 	return (
-		<DialogPrimitive.Trigger
-			data-slot="create-cost-unit-sheet-trigger"
-			handle={handle}
-			render={<Button {...props} />}
-		/>
-	);
-}
+		<DialogPrimitive.Root {...props} handle={handle}>
+			{({ payload }) => (
+				<SheetContent className={"data-nested-dialog-open:blur-xs"}>
+					<SheetHeader>
+						<SheetTitle>Edit cost unit</SheetTitle>
+					</SheetHeader>
 
-function CreateCostUnitSheet({
-	handle,
-	...props
-}: Omit<React.ComponentProps<typeof Sheet>, "handle"> & WithHandle) {
-	return (
-		<DialogPrimitive.Root handle={handle} {...props}>
-			<SheetContent
-				className={
-					"data-nested-dialog-open:-translate-x-6 data-nested-dialog-open:scale-98"
-				}
-			>
-				<SheetHeader>
-					<SheetTitle>Neue Kostenstelle</SheetTitle>
-				</SheetHeader>
-
-				<AsyncBoundary
-					pending={<CreateCostUnitFormSkeleton />}
-					rejected={({ error, retry }) => (
-						<CreateCostUnitFormError error={error} retry={retry} />
-					)}
-				>
-					<CreateCostUnitFormConnected handle={handle} />
-				</AsyncBoundary>
-			</SheetContent>
+					{payload ? (
+						<AsyncBoundary
+							// SECTION 3 — keying on the entity id remounts queries + form
+							// when a different row opens the sheet. Fresh seed every time;
+							// you never manually reset form state.
+							key={payload.id}
+							pending={<SheetFormSkeleton fieldCount={COST_UNIT_FORM_FIELD_COUNT} />}
+							rejected={({ error, retry }) => (
+								<SheetFormError error={error} retry={retry} />
+							)}
+						>
+							<UpdateCostUnitFormConnected costUnitId={payload.id} handle={handle} />
+						</AsyncBoundary>
+					) : null}
+				</SheetContent>
+			)}
 		</DialogPrimitive.Root>
 	);
 }
 
-function CreateCostUnitFormConnected({ handle }: WithHandle) {
+type UpdateCostUnitFormProps = {
+	defaultValues: UpdateCostUnitFormValues;
+	submitLabel: string;
+	groups: CostUnitGroup[];
+	canDelete?: boolean;
+
+	onSubmit: (values: UpdateCostUnitFormValues) => Promise<void> | void;
+	onDelete: () => Promise<void> | void;
+};
+
+function UpdateCostUnitFormConnected({
+	costUnitId,
+	handle,
+}: {
+	costUnitId: string;
+	handle: UpdateCostUnitHandle;
+}) {
 	const utils = api.useUtils();
 
-	const [{ data: groups }] = useSuspenseQueries({
-		queries: [utils.costUnit.listGroups.queryOptions()],
+	const [{ data: groups }, { data: costUnit }] = useSuspenseQueries({
+		queries: [
+			utils.costUnit.listGroups.queryOptions(),
+			utils.costUnit.getById.queryOptions({ id: costUnitId }),
+		],
 	});
 
-	const create = api.costUnit.create.useMutation({
+	const update = api.costUnit.update.useMutation({
 		onSuccess: (value) => {
-			toast.success("Kostenstelle wurde erfolgreich erstellt", {
+			toast.success("Kostenstelle wurde erfolgreich aktualisiert", {
 				description: `${value.tag} • ${value.title}`,
 			});
 			handle.close();
 		},
 		onError: (error) => {
-			toast.error("Fehler beim Erstellen der Kostenstelle", {
+			toast.error("Fehler beim Aktualisieren der Kostenstelle", {
+				description: error.message ?? "Ein unbekannter Fehler ist aufgetreten",
+			});
+		},
+	});
+
+	const deleteCostUnit = api.costUnit.delete.useMutation({
+		onSuccess: (value) => {
+			toast.success("Kostenstelle wurde erfolgreich gelöscht", {
+				description: `${value.tag} • ${value.title}`,
+			});
+			handle.close();
+		},
+		onError: (error) => {
+			toast.error("Fehler beim Löschen der Kostenstelle", {
 				description: error.message ?? "Ein unbekannter Fehler ist aufgetreten",
 			});
 		},
 	});
 
 	return (
-		<CreateCostUnitForm
-			groups={groups}
-			onSubmit={async (values) => {
-				await create.mutateAsync(values);
+		<UpdateCostUnitForm
+			canDelete={costUnit._count.reports === 0}
+			defaultValues={{
+				...costUnit,
+				costUnitGroupId: costUnit.costUnitGroupId ?? NO_COST_UNIT_GROUP,
 			}}
+			groups={groups}
+			onDelete={async () => {
+				await deleteCostUnit.mutateAsync({
+					id: costUnitId,
+				});
+			}}
+			onSubmit={async (values) => {
+				await update.mutateAsync(values);
+			}}
+			submitLabel="Update"
 		/>
 	);
 }
 
-type CreateCostUnitFormProps = {
-	onSubmit: (values: CreateCostUnitFormValues) => Promise<void>;
-	groups: CostUnitGroup[];
-};
+function UpdateCostUnitForm({
+	defaultValues,
+	submitLabel,
+	canDelete,
+	groups,
 
-function CreateCostUnitForm({ onSubmit, groups }: CreateCostUnitFormProps) {
-	const createGroupHandleRef = React.useRef<CreateCostUnitGroupHandle | null>(
-		null,
-	);
-	if (!createGroupHandleRef.current)
-		createGroupHandleRef.current = createCreateCostUnitHandle();
-	const createGroupHandle = createGroupHandleRef.current;
+	onSubmit,
+	onDelete,
+}: UpdateCostUnitFormProps) {
+	const deleteHandleRef = React.useRef<ReturnType<
+		typeof AlertDialogPrimitive.createHandle
+	> | null>(null);
+	if (!deleteHandleRef.current)
+		deleteHandleRef.current = AlertDialogPrimitive.createHandle();
+	const deleteHandle = deleteHandleRef.current;
 
 	const form = useForm({
-		defaultValues: {
-			tag: "",
-			title: "",
-			examples: [],
-			costUnitGroupId: NO_COST_UNIT_GROUP,
-		} as CreateCostUnitFormValues,
+		defaultValues,
 		validators: {
-			onSubmit: createCostUnitSchema,
+			onSubmit: updateCostUnitSchema,
 		},
 		onSubmit: async ({ value }) => {
 			await onSubmit(value);
@@ -266,16 +315,6 @@ function CreateCostUnitForm({ onSubmit, groups }: CreateCostUnitFormProps) {
 											Hilfe deinen Nutzern schneller eine passende Kostenstelle zu finden,
 											indem du sie in Gruppen sortierst.
 										</FieldDescription>
-										<div>
-											<DialogPrimitive.Trigger
-												handle={createGroupHandle}
-												render={
-													<Button className={"w-fit -translate-x-2.5"} variant={"link"}>
-														Neue Gruppe erstellen
-													</Button>
-												}
-											/>
-										</div>
 										{isInvalid && <FieldError errors={field.state.meta.errors} />}
 									</Field>
 								);
@@ -309,7 +348,33 @@ function CreateCostUnitForm({ onSubmit, groups }: CreateCostUnitFormProps) {
 						</form.Field>
 					</FieldGroup>
 				</form>
+				<div className="mt-12 flex flex-nowrap items-start justify-between gap-8">
+					<div>
+						<Label
+							className="mb-1 font-semibold text-base text-red-600"
+							htmlFor={"delete-cost-unit"}
+						>
+							Kostenstelle löschen
+						</Label>
+						<FieldDescription>
+							{canDelete
+								? "Diese Aktion kann nicht rückgängig gemacht werden. Kostenstellen können nur gelöscht werden, wenn noch kein Antrag existiert, der diese Kostenstelle verwendet."
+								: "Du kannst diese Kostenstelle nicht löschen, da es bereits Anträge gibt, die diese Kostenstelle verwenden."}
+						</FieldDescription>
+					</div>
+					<AlertDialogTrigger
+						disabled={!canDelete}
+						handle={deleteHandle}
+						id={"delete-cost-unit"}
+						render={
+							<Button type="button" variant={"destructive"}>
+								<TrashIcon /> Löschen
+							</Button>
+						}
+					/>
+				</div>
 			</SheetBody>
+
 			<SheetFooter className="flex flex-row items-center justify-end gap-4">
 				<SheetClose
 					render={
@@ -331,63 +396,35 @@ function CreateCostUnitForm({ onSubmit, groups }: CreateCostUnitFormProps) {
 							form={FORM_ID}
 							type="submit"
 						>
-							{isSubmitting ? "Saving…" : "Erstellen"}
+							{isSubmitting ? "Saving…" : submitLabel}
 						</Button>
 					)}
 				</form.Subscribe>
 			</SheetFooter>
-			<CreateCostUnitGroupSheet closeOnSuccess handle={createGroupHandle} />
-		</>
-	);
-}
-
-function CreateCostUnitFormSkeleton() {
-	return (
-		<>
-			<SheetBody>
-				<div className="space-y-5">
-					{["Title", "Price", "Category"].map((label) => (
-						<div className="space-y-2" key={label}>
-							<Skeleton className="h-4 w-20" />
-							<Skeleton className="h-10 w-full" />
-						</div>
-					))}
-				</div>
-			</SheetBody>
-			<SheetFooter>
-				<Skeleton className="h-10 w-32" />
-			</SheetFooter>
-		</>
-	);
-}
-
-export function CreateCostUnitFormError({
-	error,
-	retry,
-}: {
-	error: Error;
-	retry: () => void;
-}) {
-	return (
-		<>
-			<SheetBody>
-				<div className="space-y-3">
-					<p className="font-medium text-sm">Couldn&apos;t load the form</p>
-					<p className="text-muted-foreground text-sm">{error.message}</p>
-				</div>
-			</SheetBody>
-			<SheetFooter>
-				<Button onClick={retry} variant="outline">
-					Try again
-				</Button>
-			</SheetFooter>
+			<AlertDialog handle={deleteHandle}>
+				<AlertDialogContent>
+					<AlertDialogHeader>
+						<AlertDialogTitle>Kostenstelle löschen</AlertDialogTitle>
+						<AlertDialogDescription>
+							Bist du dir sicher, dass du diese Kostenstelle löschen möchtest? Diese
+							Aktion kann nicht rückängig gemacht werden.
+						</AlertDialogDescription>
+					</AlertDialogHeader>
+					<AlertDialogFooter>
+						<AlertDialogCancel>Abbrechen</AlertDialogCancel>
+						<AlertDialogAction
+							onClick={onDelete}
+							render={<Button variant={"destructive"}>Löschen</Button>}
+						/>
+					</AlertDialogFooter>
+				</AlertDialogContent>
+			</AlertDialog>
 		</>
 	);
 }
 
 export {
-	type CreateCostUnitHandle,
-	CreateCostUnitSheet,
-	CreateCostUnitSheetTrigger,
-	createCreateCostUnitHandle,
+	createCostUnitUpdateHandle,
+	type UpdateCostUnitHandle,
+	UpdateCostUnitSheet,
 };
